refactor(client): Simplify `context` setup.
Because `getContext` returned a self-created object, using it with mixins
required some unnatural coding - see [1].

Instead of having `getContext` self create an object, pass in a context model
to an `updateContext` function. `updateContext` sets the values that are
expected. This allows mixins to safely override `updateContext` without
resorting to terrible hacks.

[1] - https://github.com/mozilla/fxa-content-server/blob/7b31dc29dd1eef2b4b2b52d7f6370c144e507009/app/scripts/views/mixins/back-mixin.js#L30:L46